requirements.txt
streamlit>=1.40.0
sarvamai>=0.1.0
chromadb>=0.5.23
scikit-learn>=1.5.2
sentence-transformers>=3.0.1
requests>=2.32.3
python-dotenv>=1.0.1
pypdf>=5.1.0
beautifulsoup4>=4.12.3
